add test setup and move app to a separate file
