+++ auth/src/test/setup.ts
@@ -1,0 +1,24 @@
+import { MongoMemoryServer } from 'mongodb-memory-server';
+import mongoose from 'mongoose';
+import { app } from '../app';
+
+let mongoServer: MongoMemoryServer;
+
+beforeAll(async () => {
+  mongoServer = new MongoMemoryServer();
+  const mongoUri = mongoServer.getUri();
+  await mongoose.connect(mongoUri);
+});
+
+beforeEach(async () => {
+  const collections = await mongoose.connection.db.collections();
+  collections.forEach(async (collection) => {
+    await collection.deleteMany({});
+  });
+});
+
+afterAll(async () => {
+  // drop all collections
+  mongoServer.stop();
+  await mongoose.disconnect();
+});
